// src/models/detailed-sale.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Bespoked Bikes - API
 * BeSpoked Bikes is looking to create a sales tracking application.  BeSpoked is a high-end bicycle shop and each salesperson gets a commission for each bike they sell.  They are introducing a new quarterly bonus based on sales as an incentive.
 *
 * OpenAPI spec version: 1.0.11
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Customer } from './customer';
import { Product } from './product';
import { Salesperson } from './salesperson';
/**
 * 
 * @export
 * @interface DetailedSale
 */
export interface DetailedSale {
    /**
     * 
     * @type {number}
     * @memberof DetailedSale
     */
    id?: number;
    /**
     * 
     * @type {Product}
     * @memberof DetailedSale
     */
    product?: Product;
    /**
     * 
     * @type {Customer}
     * @memberof DetailedSale
     */
    customer?: Customer;
    /**
     * 
     * @type {Salesperson}
     * @memberof DetailedSale
     */
    salesperson?: Salesperson;
    /**
     * 
     * @type {string}
     * @memberof DetailedSale
     */
    sale_date?: string;
    /**
     * 
     * @type {number}
     * @memberof DetailedSale
     */
    quantity?: number;
}